backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
aiosqlite==0.20.0
pymysql==1.1.0
cryptography==43.0.1
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
email-validator==2.2.0
httpx==0.27.0
python-dotenv==1.0.1
yt-dlp

